use multiple target style solver arguments
In Nutils master the solvers support solving for multiple targets and returing
solutions in dictionaries. The latter comes in handy when working with multiple
arguments, as is the case in the plasticity script. This commit combines all
arguments in a `state` dictionary, which is passed to and updated by the
solvers, and passed to evaluations at samples. The target `lhs` has been
renamed to `u`, which is more in line with the other arguments in the state
dictionary.